Feature/45 priority calc

+++ application-main/src/main/java/com/honeybadgers/realtimescheduler/model/RedisTask.java
@@ -12,5 +12,5 @@ import java.io.Serializable;
 public class RedisTask implements Serializable {
 
     private String id;
-    private int priority;
+    private long priority;
 }
